data is in float

+++ bridge-to-influxdb/domotik/main.rb
@@ -15,7 +15,7 @@ begin
             _unit = elements[2]
 
             data = {
-                values: { value: message.to_i, type: _type, name: _name, unit: _unit },
+                values: { value: message.to_f, type: _type, name: _name, unit: _unit },
                 timestamp: Time.now.to_i
             }
             puts data
